requirements.txt
openai
python-dotenv
pyttx3
SpeechRecognition